% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly.r
\name{fpc_hourly}
\alias{fpc_hourly}
\title{Get hourly ladder counts from FPC}
\usage{
fpc_hourly(sdate, edate, ladder_id, species = "adultchinook")
}
\arguments{
\item{sdate}{Start date \code{character} in m-d-yyyy format. e.g., \code{"8-1-2019"}}

\item{edate}{End date \code{character} in m-d-yyyy format}

\item{ladder_id}{BON \code{1:2}, TDA \code{3:4}, JDA \code{5:6}, MCN \code{7:8}, IHR \code{9:10}, LMN \code{11:12}, LGS \code{13}, LGR \code{14}}

\item{species}{Species for which to get counts. e.g., \code{"adultchinook"} or \code{"jackchinook"}}
}
\value{

}
\description{
Downloads hourly counts of one species in one ladder at a mainstem Columbia or Snake River dam.
}
\examples{
\dontrun{
fpc_hourly(sdate="8-1-2019", edate="10-1-2019", ladder_id=1, species="adultchinook")
fpc_hourly(sdate="8-1-2019", edate="10-1-2019", ladder_id=1, species="jackchinook")
}
}
