% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laddersplit.r
\name{fpc_laddersplit}
\alias{fpc_laddersplit}
\title{Get laddersplit count report from FPC}
\usage{
fpc_laddersplit(sdate, edate, rpt = c("salmon", "lamprey"))
}
\arguments{
\item{sdate}{Start date for ladder counts}

\item{edate}{End date for ladder counts}

\item{rpt}{Report to download, either \code{"salmon"} (anadromous salmonids) or \code{"lamprey"} (all other species)}
}
\value{

}
\description{
Downloads the daily ladder split report for mainstem Columbia and Snake River dams.
}
\examples{
\dontrun{
get_laddersplit(sdate="1-1-2019", edate="12-31-2019", rpt="salmon")
}
}
