% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsum.r
\name{get_hourly}
\alias{get_hourly}
\title{Get hourly ladder counts}
\usage{
get_hourly(
  sdate = "8-1-2019",
  edate = "10-1-2019",
  ladder_id,
  species = "adultchinook"
)
}
\arguments{
\item{sdate}{Start date}

\item{edate}{End date}

\item{ladder_id}{Ladders BON 1:2, TDA 3:4, JDA 5:6, MCN 7:8, IHR 9:10, LMN 11:12, LGS 13, LGR 14}

\item{species}{Species to get counts for}
}
\value{

}
\description{
Get hourly ladder counts
}
\examples{
\dontrun{
get_hourly(sdate="8-1-2019", edate="10-1-2019", ladder_id=1, species="adultchinook")
get_hourly(sdate="8-1-2019", edate="10-1-2019", ladder_id=1, species="jackchinook")
}
}
