% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsum.r
\name{get_laddersplit}
\alias{get_laddersplit}
\title{Download laddersplit count report}
\usage{
get_laddersplit(sdate, edate, rpt = c("salmon", "lamprey"))
}
\arguments{
\item{sdate}{Start date}

\item{edate}{End date}

\item{rpt}{Report to download, either "salmon" (anadromous salmonids) or "lamprey" (all other species)}
}
\value{

}
\description{
Download laddersplit count report
}
\examples{
\dontrun{
get_laddersplit(sdate="1-1-2019", edate="12-31-2019", rpt="salmon")
}
}
