% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsum.r
\name{get_runsum}
\alias{get_runsum}
\title{Get runsum report}
\usage{
get_runsum(
  dam = c("BON", "TDA", "JDA", "MCN", "IHR", "LMN", "LGS", "LGR"),
  sdate,
  edate,
  rpt = c("salmon", "lamprey")
)
}
\arguments{
\item{dam}{Dam at which counts are desired, only available for Corps dams in this query.Dams: BON, TDA, JDA, MCN, IHR, LMN, LGS, LGR}

\item{sdate}{Start date to pull counts. \code{character} in m-d-yyyy format.}

\item{edate}{End date to pull counts. \code{character} in m-d-yyyy format.}

\item{rpt}{Report to download, either "salmon" or "lamprey"}
}
\value{
The data frame downloaded from FPC.
}
\description{
Get runsum report
}
\examples{

\dontrun{
get_runsum(dam="BON", sdate="1-1-2019", edate="12-31-2019", rpt="salmon")
}
}
